move journal fetcher
