+++ src/main/java/org/jabref/gui/fieldeditors/journalinfo/JournalInfoViewModel.java
@@ -14,7 +14,7 @@ import javafx.util.Pair;
 import org.jabref.gui.AbstractViewModel;
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.journals.JournalInformation;
-import org.jabref.logic.journals.JournalInformationFetcher;
+import org.jabref.logic.importer.fetcher.JournalInformationFetcher;
 
 public class JournalInfoViewModel extends AbstractViewModel {
     private final ReadOnlyStringWrapper title = new ReadOnlyStringWrapper();
